springboot-jpa/src/test/java/io/ensueno/repository/MemberRepositoryTest.java
package io.ensueno.repository;

import io.ensueno.Entity.Member;
import org.assertj.core.api.Assertions;
import org.junit.Test;
import org.junit.jupiter.api.DisplayName;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.Rollback;
import org.springframework.test.context.junit4.SpringRunner;

import javax.transaction.Transactional;

@RunWith(SpringRunner.class)
@SpringBootTest
public class MemberRepositoryTest {

    @Autowired MemberRepository memberRepository;

    @Test
    @Transactional
    @Rollback(false)
    public void testMember() throws Exception {
        //given
        Member member = new Member();
        member.setUsername("memberA");
        //when
        Long saveId = memberRepository.save(member);
        Member findMember = memberRepository.find(saveId);

        //then
        Assertions.assertThat(findMember.getId()).isEqualTo(member.getId());
        Assertions.assertThat(findMember.getUsername()).isEqualTo(member.getUsername());

        // findMember와 member는 같은 트랜잭션 안에서는 영속성 컨텍스트가 id 값이 같을 경우 같은 identity로 식별 한다.
        // 1차 캐시에서 영속성 컨텍스에 관리되고 있던 멤버가 달려 오게 된다.
        Assertions.assertThat(findMember).isEqualTo(member);
    }

}